Lesson 70 Add custom restaurant authorization service and handlers
Introduce IRestaurantAuthorizationService and implementation to enforce per-user and per-role authorization for restaurant operations. Add ResourceOperation enum and ForbidException. Update command handlers to check authorization before update/delete actions. Enhance ErrorHandlingMiddleware to handle forbidden access. Register new service in DI. Update HTTP test file for new auth logic.

+++ Restaurants.Application/Dishes/Commands/DeleteDish/DeleteDishForRestaurantCommandHandler.cs
@@ -3,13 +3,15 @@ using MediatR;
 using Microsoft.Extensions.Logging;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDish;
 
 internal class DeleteDishForRestaurantCommandHandler(ILogger<DeleteDishForRestaurantCommandHandler> logger,
 IRestaurantsRepository restaurantRepository,
-IDishesRepository dishesRepository) : IRequestHandler<DeleteDishForRestaurantCommand>
+IDishesRepository dishesRepository
+	, IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteDishForRestaurantCommand>
 {
 	public async Task Handle(DeleteDishForRestaurantCommand request, CancellationToken cancellationToken)
 	{
@@ -21,6 +23,9 @@ IDishesRepository dishesRepository) : IRequestHandler<DeleteDishForRestaurantCom
 		var dish = restaurant.Dishes.FirstOrDefault(p => p.Id == request.DishId);
 		if (dish == null ) throw new NotFoundException(nameof(Dish), request.DishId.ToString());
 
-		await dishesRepository.DeleteByIdAsync(dish);
+        if (!restaurantAuthorizationService.Authorize(restaurant, Domain.Constants.ResourceOperation.Update))
+            throw new ForbidException();
+
+        await dishesRepository.DeleteByIdAsync(dish);
 	}
 }
