+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -2,12 +2,14 @@
 using Microsoft.Extensions.Logging;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
 
 public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger,
-	IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+	IRestaurantsRepository restaurantsRepository,
+	IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
 {
 	public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
 	{
@@ -17,6 +19,9 @@ public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandl
 		{
 			throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
 		}
+
+		if (!restaurantAuthorizationService.Authorize(restaurant, Domain.Constants.ResourceOperation.Delete))
+			throw new ForbidException();
 
 		await restaurantsRepository.Delete(restaurant);
 	}
